requirements.txt
mysql-connector-python
python-multipart
python-dateutil
fastapi
uvicorn
loguru
python-dotenv
